Refactor node-mocks commands
