+++ packages/ui/package.json
@@ -9,9 +9,9 @@
     "lint:eslint": "eslint './{src,test}/**/*.{ts,tsx}'",
     "lint:fix": "yarn lint:eslint --fix && yarn lint:prettier --write --loglevel warn",
     "lint:prettier": "yarn prettier './{src,test}/**/*.{ts,tsx,html}'",
-    "node-mocks": "ts-node dev/node/index.ts",
+    "node-mocks": "ts-node dev/node-mocks/index.ts",
     "queries:generate": "graphql-codegen --config codegen.config.yml && yarn lint:fix",
-    "query-node-mocks": "ts-node dev/node-mocks/generateMocks.ts",
+    "query-node-mocks": "ts-node dev/query-node-mocks/generateMocks.ts",
     "start": "NODE_OPTIONS=--max_old_space_size=8192 webpack serve --mode development --progress",
     "storybook": "start-storybook -p 6006 -s public --no-manager-cache",
     "test": "jest"
